Data validation

+++ api/Controllers/GameController.cs
@@ -30,8 +30,10 @@ namespace api.Controllers
             return Ok(gamesDto);
         }
 
-        [HttpGet("{id}")]
+        [HttpGet("{id:int}")]
         public async Task<IActionResult> GetById([FromRoute] int id) {
+            if (!ModelState.IsValid) return BadRequest(ModelState);
+
             var game = await _gameRepository.GetGameByIdAsync(id);
 
             if (game == null) return NotFound();
@@ -42,6 +44,8 @@ namespace api.Controllers
         [HttpPost]
         [Route("addGame")]
         public async Task<IActionResult> CreateNewGame([FromBody] CreateGameRequestDto requestDto) {
+            if (!ModelState.IsValid) return BadRequest(ModelState);
+
             var gameModel = requestDto.ToGameFromCreateDtoRequest();
 
             await _gameRepository.AddGameAsync(gameModel);
@@ -51,8 +55,10 @@ namespace api.Controllers
         }
 
         [HttpPut]
-        [Route("update/{id}")]
+        [Route("update/{id:int}")]
         public async Task<IActionResult> Update([FromRoute] int id , [FromBody] UpdateGameRequestDto updateRequestDto) {
+            if (!ModelState.IsValid) return BadRequest(ModelState);
+
             var gameModel = await _gameRepository.UpdateGameAsync(id , updateRequestDto);
 
             if (gameModel == null) return NotFound();
@@ -61,8 +67,10 @@ namespace api.Controllers
         }
 
         [HttpDelete]
-        [Route("delete/{id}")]
+        [Route("delete/{id:int}")]
         public async Task<IActionResult> Delete([FromRoute] int id) {
+            if (!ModelState.IsValid) return BadRequest(ModelState);
+            
             var gameModel = await _gameRepository.DeleteGameByIdAsync(id);
 
             if (gameModel == null) return NotFound();
